Major Update
- Files Merged Successfully
- Notifications now working in Intervals
- Removal of Spoilage Reports in Commissary Notifications
- Generation of Spoilage Report Summary for Cafe and Intern
- Adjustment of Commissary Stock and Cafe Stock column data types in the DB (Updated to Double)

+++ BE/ims_backend/ims/models.py
@@ -5,8 +5,8 @@ from django.contrib.auth.models import User
 # Create your models here.
 class Item(models.Model):
     item_name = models.CharField(max_length=64)
-    commissary_stock = models.IntegerField()
-    cafe_stock = models.IntegerField(default=0)
+    commissary_stock = models.FloatField()
+    cafe_stock = models.FloatField(default=0)
     um = models.CharField(max_length=16)
     category = models.CharField(max_length=64)
